empezando el paso 2

+++ modelo/ventilacion.sql
@@ -6,7 +6,7 @@ use ventilacion;
 create table pacientes(
 	id_paciente int unsigned not null auto_increment,
 	nombre varchar(100) not null default "desconocido",
-	peso smallint not null default 0,
+	peso float(6,2) not null default 0,
 	estatura smallint not null default 0,
 	genero varchar(3) not null default "I",
 	frecuenciaRespiratoria smallint not null default 0,
@@ -14,8 +14,8 @@ create table pacientes(
 	presionPico smallint not null default 0,
 	presionMeseta smallint not null default 0,
 	peep smallint not null default 0,
-	fio2 smallint not null default 0,
-	poderMecanico float(6,2) not null default 0,
+	fio2 float(8,2) not null default 0,
+	poderMecanico float(8,2) not null default 0,
 	status tinyint not null default 1,
 	PRIMARY KEY (id_paciente)
 )engine=innodb, charset=utf8;
